+++ ncalendar/src/main/java/com/necer/listener/OnCalendarScrollingListener.java
@@ -2,5 +2,5 @@ package com.necer.listener;
 
 //日历 月 周 拉伸 状态滑动监听
 public interface OnCalendarScrollingListener {
-    void onCalendarScrolling(int dy);
+    void onCalendarScrolling(float dy);
 }
